Commented out the declaration of some undefined functions in beat.h, and the unused function eof_change_bpm().
Made a very minor optimization in eof_calculate_beats() that finds the index of the last beat in the beat[] array more efficiently

Made a very minor optimization in eof_realign_beats() that finds the number of beats between the previous and next anchor more efficiently.  I applied the same optimization in eof_recalculate_beats() a couple of times

Added bounds checking for the functions that indexed into eof_song->beat[] without verifying the index against eof_song->beats first

Added a comment line to describe the behavior of the functions declared in beat.h.