Works on issue 117.
eof_note_draw_quick() was implemented in the edit window rendering.  Now, EOF stops trying to draw notes when it knows the rest of them will render off-screen.

eof_note_draw_3d() implemented skipping the rendering if it would be outside bounds, but it didn't return a clipping status, so I modified it to behave like eof_note_tail_draw_3d().

eof_render_3d_window() was slightly optimized to combine two loops and remove some debugging code.  Since eof_note_tail_draw_3d() already implemented a method of skipping rendering based on certain time position bounds, I put it to use to break out of rendering early if notes are not visible. 3D notes are rendered last to first so that the nearest notes are in front, so the stop rendering logic is opposite:  The rendering loop breaks if notes are before the visible area.